Rename teampage_infoblock.js to info_block.js
