+++ src/components/team/info_block.js
